Isolate settings in a separate module
